requirements_v5_v6_v7.txt
# Requirements para V5, V6 e V7
# Para instalar: pip install -r requirements_v5_v6_v7.txt

# Modelos de Machine Learning
lightgbm>=4.1.0        # V5 - LightGBM
catboost>=1.2.2        # V6 - CatBoost
xgboost>=2.0.0         # Já usado no V4, necessário para V7

# Bibliotecas base (já devem estar instaladas)
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Google Cloud (já instalado)
google-cloud-bigquery>=3.11.0

# Para salvar meta-learner do V7
pickle5>=0.0.11; python_version < '3.8'
